Update README and refactor configuration files for Jest to ts
