+++ src/app/api/status/route.ts
@@ -3,3 +3,4 @@ import { NextResponse } from 'next/server';
 export async function GET() {
   return NextResponse.json({ status: 'ok' });
 }
+
