+++ src/app/layout.tsx
@@ -1,8 +1,9 @@
 import "./globals.css";
+import React from "react";
 
 export default function RootLayout({
-  children,
-}: Readonly<{
+                                     children,
+                                   }: Readonly<{
   children: React.ReactNode;
 }>) {
   return (
