+++ src/app/page.tsx
@@ -2,7 +2,8 @@ import Image from "next/image";
 
 export default function Home() {
   return (
-    <div className="grid grid-rows-[20px_1fr_20px] items-center justify-items-center min-h-screen p-8 pb-20 gap-16 sm:p-20">
+    <div
+      className="grid grid-rows-[20px_1fr_20px] items-center justify-items-center min-h-screen p-8 pb-20 gap-16 sm:p-20">
       <h1 className="text-4xl font-bold text-center">Hello, World!</h1>
       <Image
         src="/saturne.png"
@@ -13,3 +14,4 @@ export default function Home() {
     </div>
   );
 }
+
