+++ tailwind.config.ts
@@ -16,3 +16,4 @@ export default {
   },
   plugins: [],
 } satisfies Config;
+
